Deimos Bazzar Sell for sigil
Added multi user selling logic. Calcuates backpack space + gold capacity. If backpack is close to full, it goes and sells to the bazzar. If users gold is full, it doesnt sell and just keeps farming items forever. It only sells when the user money is less than 5% of max and if the backpack is full